BUG 2112: Workaround for missing flows in POC demo.
When the POC demo is run, one of the vSwitches occasionally
is missing some of its flows, due to a problem in the openflowplugin.
This commit introduces a workaround to avoid this problem and allow
the demo to work.

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>
Change-Id: I0f7dca025b30b032d68ee1bd83d0e82966efa2df
(cherry picked from commit d24e285cb82236eeddd72e33fbe3a04acd600de1)