Fix build by referencing correct dependencies.
This patch fixes a broken build by pulling in the
correct version of PAX exam. It also migrates the
karaf container from 3.0.1 to 3.0.2.  This fix should
eventually be replaced by a patch that converts to the
pom/project structure, as per the ODL "best practices"
that is defined in the archetype.

Change-Id: Ia626643c2cdb2bedc2b9df08cc267a8aa9902ad1
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>