Some thoughts for SFC Integration.
This patch is just meant to kick off some initial thinking
about what an SFC/GBP integration might look like. It creates
a manager that listens for ActionInstance updates, then gets
the ActionDefinition that it references to see if the action
name is meant to direct to a Service Function Chain, and if
so, uses the chain name provided in the ActionInstance to
make the call into SFC to get the RendereredServicePath, by
name.

Change-Id: I8fbc5082c697d96d737b784a431cc484534ce715
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>