Scope change for OfTable.
The OfTable class provides protected acces to OpenFlow Overlay
renderer classes for the Flow Tables (e.g. SourceMapper,
DestinationMapper, et. al.). This limits access to just
the Flow Table classes themselves. This patch moves that
to a separate class with private members and public getters
so that the state can be accessed by any class that has the
new object (e.g. Action classes will use this in future patches).

Change-Id: I21aefa959c97453cdfa8f7f5b668f468dfd7aa5d
Signed-off-by: Thomas Bachman <tbachman@yahoo.com>