Bug 3460 removed ip prefix from DHCP/DNS flows
Removes ip-prefix of EIC from and DHCP and DNS flows.
Only traffic based on destination port of these service is allowed.
IP prefix of EIC for ICMP is still part of flow match.

Removed comparison of flow priority in FlowEquivalence - duplicit flows are ignored

THIS COMMIT HAS IMPACT ON FLOW COUNT IN TABLE 4

Change-Id: Id5dec7cdc512ed92e479992102696970dc7784ca
Signed-off-by: Martin Sunal <msunal@cisco.com>