Introduced Renderer Manager
Renderer manager dispatches policy configuration among endoints
to renderers.
Renderer manager gets data notifications about:
- Endoints (base-endpoint.yang)
- EndpointLocations (base-endpoint.yang)
- ResolvedPolicies (resolved-policy.yang)
- Forwarding (forwarding.yang)

Based on data from notifications it creates a configuration for renderers.
Configuration is stored to CONF DS as Renderers (renderer.yang)

Configuration is sign with version which is incremented by every change.
All rendrers are suposed to be on the same version. Renderer manager waits
for all renderers to respond with version update in OPER DS. After a version
of every renderer in OPER DS is same as version in CONF DS, renderer manager
moves to the next configuration with incremented version.

Change-Id: I475b357d5d60444dc359bc400c567d9d0e36238f
Signed-off-by: Martin Sunal <msunal@cisco.com>