Moved l2-l3 forwarding to domain-extensions
domain-extensions is location for all domain specific
extensions of forwarding and endpoints.
Purpose of puting a domain specific extension
to domain-extensions is that multiple renderers
can depends on it.

The first domain specific extension is l2-l3 (layer 2 and layer 3).
The l2-l3 extension extends forwarding with
L2 and L3 specific contexts and network domains.
It is expected to have L2, L3 augmentations of an endpoint here.

Change-Id: I85e8bc1040acc7ce95a3b5bb9f65ad0e183e12fb
Signed-off-by: Martin Sunal <msunal@cisco.com>