Update of mapping neutron port to vpp endpoint
Ports are mapped to vpp endpoints only if:
  - device owner is a compute or dhcp node
  - vif_type is vhostuser

Socket ID of vpp endpoints consists of:
  - vhostSocketPath + vhostSocketPrefix + portUUID

Incorporated the vpp-node property in vpp endpoints
which provides basic info about location of an endpoint.
Addition of dependency on network topology was needed.

Removing original vpp endpoints as a first step on neutron
port update events in case that updated ports are not
validated as vpp endpoints anymore.

Change-Id: I02ebf69c261a9103a2f62c06520e4bc0ca66f57b
Signed-off-by: Tomas Cechvala <tcechval@cisco.com>