Added check for config change in RendererManager
RendererManager compares current state with last configured state.
Only if these two states are differerent a configuration for renderers
is created.
This check is needed because it happens that a producer rewrites data
in datastare with the same data and it trigger DataTreeChangeEvent
so RendererManager thinks that data was changed but it was not.

Change-Id: Idc145f37351899e1fa9eb70f62a00c8c7d54ebaf
Signed-off-by: Martin Sunal <msunal@cisco.com>