Netconf transactions synchronization GBP part
adding mechanism to sync transactions between GBP and VBD
this is used as a workaround for bug https://bugs.opendaylight.org/show_bug.cgi?id=7918
In order for this workaround to work GBP uses Reentrant lock from VBD to synchronize transactions.
It is needed to use netconfSyncedWrite and netconfSyncedDelete from GbpNetconfTransaction
for all write and delete operations to netconf device.

Change-Id: I4f74c376a06adb5aec058710ccbb8ed069103483
Signed-off-by: Michal Cmarada <mcmarada@cisco.com>