Add config service for loading cfg file
The existing way of using configs in vpp-renderer is to use blueprint
and pass the config as bean constructor parameter. The problem with
that is, when there are a lot of configs in the cfg file, we can
possibly pass all of them as bean constructor parameter. Hence, this
patch implements a Config Service which dynamically reads the config
and incase there is any update of config on runtime, the update is
propagated to the service too.

Change-Id: I015f61040a23d472db2a6e9eb23de94625fccb3c
Signed-off-by: Shakib Ahmed <sheikahm@cisco.com>