Support for metadata in HA
Metadata service in Openstack is reachable through every DHCP
port that is created in HA scenarios. This requires severe
changes in current forwarding design. First, Metadata is
registered as an endpoint, and since in HA it is reachable
through multiple local ports, it is assigned relative location
that is available in current models.

TODO extend forwarding APIs for this use case

Change-Id: I097f66eb04da8419a5d6dd551ad4c28f6185efb4
Signed-off-by: Tomas Cechvala <tcechval@cisco.com>