Fix for SfcProviderRpc and IpSgtDistribution failures
SFC removed static implementation of SfcProviderRpc. This patch creates an
an instance of SfcProviderRpc in SfcManager to be used instead of static
implementation.
IpSgtDistribution tests were failing because of errors in pom files. Once
BindingNormalizedNodeCodec was refactored in MDSal tests failed because
of class not found exception. Pom files were fixed to address this issue.


Change-Id: If323f84f98995751b755b0aab3be04e88fc4f24e
Signed-off-by: Michal Cmarada <mcmarada@cisco.com>