SNAT implementation for DVR
The feature is implemented in VPP renderer.
There are two implementation of SNAT:
 - centralized
 - DVR

In centralized scenario, NAT is configured on one dedicated node.
The node is the one where BVI interface is configured. All static
entries and dynamic entries are on the same node.

In DVR scenario, NAT is configured on every node where endpoints
resode. Static entries on a node are significant only to local
endpoints. There is no support for dynamic NAT yet.

Change-Id: If3c0e80b0fb1c98b85a8c8d7cbd7c25000d33238
Signed-off-by: Tomas Cechvala <tcechval@cisco.com>
Signed-off-by: Michal Cmarada <mcmarada@cisco.com>